Font-weight of Prices Changed
From the design on Figma, the prices and second row of the table are not bold. Because we've set a general style to bold all text in our internal css, these prices are also bold. 

I then give each part an inline css to change their font-weight to normal, which tally with what is on Figma